requirements.txt
# pip3 install -r requirements.txt
fastapi
uvicorn
pydantic_settings
python-dotenv
pymongo